+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### [13.2.1](https://kaos.sh/ek/13.2.1)
+
+- `[terminal/input]` Added `NewLine` flag
+- `[sliceutil]` Methods `Copy`, `IsEqual`, `Index`, `Contains`, and `Deduplicate` marked as deprecated
+- `[terminal/input]` Improved TMUX support
+
+
 ### [13.2.0](https://kaos.sh/ek/13.2.0)
 
 - `[errutil]` Added method `Wrap`